consommateur/requirements.txt
Flask
pymongo
bcrypt
plotly
requests
kafka-python

